+++ migrations/2018-07-14-112345_create_users/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE user (
-  telegram_id   INT         PRIMARY KEY,
+  telegram_id   BIGINT      PRIMARY KEY,
   name          TEXT        DEFAULT NULL,
   created_at    DATETIME    NOT NULL        DEFAULT CURRENT_TIMESTAMP,
   updated_at    DATETIME    NOT NULL        DEFAULT CURRENT_TIMESTAMP
