+++ migrations/2018-07-14-112350_chat_user_stats/up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE chat_user_stats (
   id            INT         AUTO_INCREMENT  PRIMARY KEY,
-  chat_id       INT         NOT NULL,
-  user_id       INT         NOT NULL,
+  chat_id       BIGINT      NOT NULL,
+  user_id       BIGINT      NOT NULL,
   messages      INT         NOT NULL        DEFAULT 0,
   created_at    DATETIME    NOT NULL        DEFAULT CURRENT_TIMESTAMP,
   updated_at    DATETIME    NOT NULL        DEFAULT CURRENT_TIMESTAMP,
